+++ src/Fluxor.Extensions.StoreLogger/LocalStorageMiddlewareExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using Fluxor.DependencyInjection;
+using Microsoft.Extensions.DependencyInjection;
 
 namespace Fluxor.Extensions.StoreLogger
 {
@@ -7,6 +8,7 @@ namespace Fluxor.Extensions.StoreLogger
         public static Options AddStoreLoggerMiddleware(
             this Options options, StoreLoggerOptions loggerOptions)
         {
+            options.Services.AddSingleton<StoreLoggerOptions>((prov) => loggerOptions);
             options = options.AddMiddleware<FluxorStoreLoggerMiddleware>();
             return options;
         }